requirements.txt
altgraph==0.16.1
astroid==2.2.5
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
backports.csv==1.0.7
bleach==3.1.0
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.2
django-crispy-forms==1.7.2
django-import-export==1.2.0
django-phone-field==1.7.1
django-phonenumber-field==3.0.1
entrypoints==0.3
et-xmlfile==1.0.1
future==0.17.1
imageio==2.5.0
imageio-ffmpeg==0.3.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.20
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.1
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
mpmath==1.1.0
nbconvert==5.6.0
nbformat==4.4.0
nose==1.3.7
notebook==6.0.1
numpy==1.17.0
odfpy==1.4.0
opencv-python==4.1.0.25
openpyxl==2.6.3
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pefile==2019.4.18
phonenumbers==8.10.18
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
Pygments==2.4.2
PyInstaller==3.5
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.1
pywin32==225
pywin32-ctypes==0.2.0
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
scipy==1.3.1
Send2Trash==1.5.0
six==1.12.0
sqlparse==0.3.0
sympy==1.4
tablib==0.13.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
typed-ast==1.4.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.1
xlrd==1.2.0
xlwt==1.3.0